requirement.txt
rasa==3.6.14
langchain==0.1.13
langchain-google-genai==0.0.9
chromadb==0.4.24
pymupdf==1.23.26
langchain-community==0.0.29
tqdm==4.66.2
# Webhook server dependencies
Flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
google-cloud-translate==3.11.2
google-cloud-speech==2.24.0

